+++ src/main/java/com/myprojects/invoices_frontend/services/ProductsService.java
@@ -2,11 +2,13 @@ package com.myprojects.invoices_frontend.services;
 
 import com.myprojects.invoices_frontend.clients.ProductsClient;
 import com.myprojects.invoices_frontend.domain.Products;
+import com.myprojects.invoices_frontend.layout.dialogboxes.ShowNotification;
 import com.myprojects.invoices_frontend.mappers.ProductsMapper;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -15,6 +17,7 @@ public class ProductsService {
     private List<Products> productsList;
     private static ProductsClient productsClient;
     private static ProductsService productsService;
+    private static InvoicesService invoicesService = InvoicesService.getInstance();
     private static ProductsMapper productsMapper = new ProductsMapper();
 
     public ProductsService(ProductsClient productsClient) {
@@ -48,6 +51,14 @@ public class ProductsService {
     }
 
     public void deleteProduct(@NotNull Products product) {
-        productsClient.deleteProduct(productsMapper.mapToProductDto(product));
+        if(invoicesService.getInvoicesList().stream()
+                .flatMap(i -> i.getProductsList().stream())
+                .noneMatch(p -> Objects.equals(p.getId(), product.getId()))) {
+            productsClient.deleteProduct(productsMapper.mapToProductDto(product));
+        } else {
+            ShowNotification cantBeDeleted = new ShowNotification("Nie można usunąć produktu "
+                    + product.getName() + ". Jest z nim powiązana wystawiona faktura!", 5000);
+            cantBeDeleted.show();
+        }
     }
 }
